Support asynchronous rowGetter (#302)
The `rowGetter` gets called with a second, optional parameter `end`. If
its return value has the property `then`, it get treated as a
Promise. Otherwise `rowGetter` might return an array with all rows
(`begin <= idx && idx < end`) or a single row at `begin` (the old
behaviour).

While the promise is not fulfilled, a blank content of the table gets
shown or an optional panel, provided by the property `loadingView` of
ReactDataGrid. This view gets the properties `loading` and `width` and
might return a `<div hidden=!loading>` or `undefined`.