Update InteractionMasks.js
When scrolling using tab, if you scroll past the current view and the try to use shift tab to move back to the first column, then the grid will not scroll back to the first column. This fix allows you to tab off of the current view and then shift tab back to the first column.