Allow rowClass to send row index,
this will allow classes to be added positionally

this change still supports adding classes by looking at the value with no breaking changes..

Now with positional support, if I want a css class applied to the last row in the table. Or the first row in the table.
Currently you have to use a css pseudo class like .rdg-row:nth-child(14) to make the 14th row have a certain css behavior. An example of this has been added to the AllFeatures demo.